refactor(folderStructure): 가독성 및 유지보수를 위한 폴더구조 변경
