+++ src/inputHandlers/getString.js
